requirements.txt
streamlit
langchain
langchain-openai
langchain-community
faiss-cpu
openai
tiktoken
python-dotenv